Fix streak extension across script runs.
When we persist datetime.datetime objects to the database they're stored
as ISO 8601 timestamps, but we don't (can't) automatically convert them
back on load, since they're in an opaque data structure (JSON).

It would be better not to use JSON to store these streaks, but for now
just bandaid over the problem by casting the timestamps back to
datetime.datetime.