Speed up race/role/char/god highscore searching.
You need a composite index across score and the column you're filtering
by to prevent needing to sort a lot of games in a temporary table. This
unfortunately adds a lot of index junk, but is the simplest approach for
now.

Longer term, we should probably normalise these columns out of the score
table which should let us perform a fast join from a smaller
heavily-indexed table into the full score table.