Handle inconsistent name capitalisation in streaks.
Previously, when players changed capitalisation of their name during a
streak (eg moving from playername on CAO to PlayerName on CDO) the
streak would think these two names are different people.

More work has to be done so player names are fully case-insensitive, but
this works for streaks at least.