More work normalising database structure.
The games table is about as normalised as we'd want. Loading data into
the database works, but we need to update the rest of model.py and then
scoring.py/write_website.py.

Most likely we need a new way to provide game data to the rest of the
application that totally hides the underlying database structure. Might
be time to move to sqlalchemy ORM???