Improve handling log import errors.
Generally errors during log import are because two processes tried to
create the same record at the same time (eg creating the same account
row). On the side that fails, just retry the transaction three times and
only raise an error if it still fails.

Also remove some error handling code for duplicate keys/etc. This will
need to be re-added in future, but it was MySQL-specific and poorly
written anyway.