Initial work to improve handle_player_streak perf.
Almost all its time is spent in model.get_player_streak. No obvious
optimisation route with Sqlite, but with Postgres it may be possible to
change the order of the .filter() clause...?? Probably not. What would
be better is getting the player's active streak in a deferred thread as
soon as we start processing the game. Anyway... this branch is a TODO.