Add player_wins_per_species/background/god funcs
These could be used as part of a move away from generating player pages
offline. Specifically, in that scenario we don't want to rely on the
player page loading every player game -- we should be able to get
aggregate stats relatively quickly even for players with tens of
thousands of games, and without needing to transfer all that data to the
client. So -- we need to generate this data on the server side directly
from database queries.

These queries work, but aren't being committed to master yet for a few
reasons:
* They're much slower than the current approach, each function call can
  take several seconds. There may be an index which could be added to
  fix this.
* Until we actually go down the path of generating pages on-demand, this
  is just duplicate work. If we're already loading every winning game
  for every player, we can derive stuff like this most efficiently from
  that data.