Use less memory generating player pages
Instead of loading every game for each player into memory (which has
issues with eg Sebi's 100k games), load only their wins (which are
only required for some functions that could be rewritten in SQL if the
memory usage / performance is still a problem).

This might slow down processing as there may be missing indexes or
effort duplication.