Tweak CSS styling
The most significant change is a reduction of body line-height to 1.3. I think this is warranted as many people have mentioned that they like how compact the old CAO scoreboard is, and that this one is too spread out.  This increases the density while still remaining pleasing to the eye.