Change formatting of unplayable wins
Don't use a <hr> because it's too large and visually complex. Instead,
show the unplayable wins directly below but in small font. The aim is to
show them for players, but to de-emphasise them, so we don't encourage
people to feel bad they aren't going back and playing old classes.